Update directory structure and package names
