+++ gradle.properties
@@ -3,7 +3,7 @@ group=org.ballerinalang
 version=1.1.0-beta.3-SNAPSHOT
 puppycrawlCheckstyleVersion=8.18
 bouncycastleVersion=1.61
-ballerinaGradlePluginVersion=0.9.1
+ballerinaGradlePluginVersion=0.10.0
 
 ballerinaLangVersion=2.0.0-beta.2.1
 stdlibTimeVersion=2.0.0-beta.2
