+++ native/src/main/java/io/ballerina/stdlib/crypto/Constants.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.ballerinalang.stdlib.crypto;
+package io.ballerina.stdlib.crypto;
 
 import io.ballerina.runtime.api.utils.StringUtils;
 import io.ballerina.runtime.api.values.BString;
